requirements.txt
requests==2.13.0
paho_mqtt==1.2
